disasm: Analyze full semantics of matched instruction
The previous expression substitution worked for the processors that
I tested with, but may not work with arbitrary definitions. Also it
was only useful in determining the values of placeholders, not as a
base for further analysis.

A disadvantage of the new approach is that it is much slower. But
that is something we can optimize later.