instr: Put consistency check for LoadedValues in standalone function
This allows it to be called on codeblock builders as well.

Verifying a builder can be done manually while debugging: checking at
every emit would slow things down. Eventually builders will produce
a code block, for which verification is run, so the value of checking
builders too is not finding more errors, but finding them earlier.