disasm: Keep track of decoded operand and argument values separately
I had both in a single data structure at first, because they are
closely related. However, there are nasty little differences in
requirements that cause problems, so it turned out to be easier
to use separate data structures.

The analysis code wants to use references within the context of
a single builder, while the operands code wants to use expressions
that can be examined outside the context of any builder.

This fixes the matching of decoded values to operands, which got
broken (again) in 764bf63f (Analyze full semantics of matched
instruction). In fact, in the original version of that commit it
worked fine, but the code builder could break and after I fixed
the way the code builder was used the operands broke.