instr: Do not remove local variables that are part of returned reference
This was broken since the overriding of the returned reference
was introduced in commit 9bd561ed (Support creating a code block
returning a reference other than 'ret'): removeRedundantNodes()
was still relying on the name 'ret' to avoid removal of returned
local variables.

I think inlineBlock() is still broken, since it not only relies
on the name 'ret' but also assumes there is at most one remaining
local variable.