instr: Remove 'name' attribute from Variable
Two different storage objects are always considered unequal; the name
does not matter. Therefore the name was mostly a hassle: there was no
requirement for it to be unique within a scope, but non-unique names
would still be confusing.

Variable storages no longer have names, but a variable's name is used
to look up a reference to the storage in a builder's namespace.