instr: Have LocalNamespace Import references unmodified
Since a SingleReference no longer belongs to a particular builder,
global references can be used as-is in a local namespace.

It is still useful to keep track of imports, to make sure that
a name that exists in the parent namespace can either be imported
or overridden, but if both are attempted in the same scope it will
be reported as an error.