instr: Replace Reference.clone() with substitute()
Since a SingleReference no longer belongs to a particular builder,
the main purpose of the clone() method is no longer to create a deep
copy, but to apply substitution functions. It was renamed to reflect
this new purpose.

I also changed the split of responsibilities between the method and
the substitution functions: the functions now take a storage and an
expression as their arguments, instead of a SingleReference and a
FixedValue. This means the code to avoid unnecessary new object
instantations and the code to apply expression substitution both
on FixedValues and storages is now inside the Reference subclasses,
avoiding duplication of such code in callers.