instr: Only rerun unused load removal if there is chance of success
The only case in which it is worth to run more than one pass is when
a removed load was itself the only reason another load's value was
considered to be in-use. We now check whether this might be the case
rather than assuming that every removed load has this potential.

There is no difference in execution time when I tested it, since the
disassembly did not trigger any unused load removals, so the function
wasn't being rerun in the first place. The unit test suite does
trigger unused load removal, so it is not a dead feature, just one
that isn't triggered yet by the disassembler in its current state.