instr: Apply load replacements in same pass that computes them
The critical realization is that expressions that precede a load
cannot use its loaded value. Therefore we get the same result by
applying the replacements gathered up until that point compared
to applying all replacements once we have the full set.

In practice the set of replacements is empty quite often, so the time
saved by doing everything in a single pass is neglible. However, this
change paves the way for future optimizations.