instr: Return final local variable values as FixedValues
Previously, those local variables that were part of the returned
value or reference had their storages inlined in the calling block
and then eliminated during later simplification. The new approach
leads to slightly simpler code, simpler unit tests and less repeated
work when a block is inlined multiple times.

It does change the behavior of the definition language slightly;
the new behavior was documented. Returning a reference to a local
variable is now effectively the same as returning its value.
I don't know of any language in which returning a reference to a
local variable is a useful feature, so I don't think users are
going to be negatively affected by this.