instr: Made 'complexity' into a property of Expression
I think it was a property originally before putting all code related
to simplification in a separate module. The problem with computing
this separately is that it is easy to forget to add support for new
subclasses, which did in fact happen.

This is part of an effort to reduce the use of 'instanceof'.