instr: Undo "treat placeholder value more like a function argument"
This reverts the main change of commit dd9e7aeb, but it is not
an exact revert.

The code simplification from that commit doesn't look like it will
last with the changes I have planned.

Meanwhile, I don't really like the new behavior conceptually:
the reason to have local variables for function arguments is to
be consistent with expectations users have from using mainstream
programming languages. With placeholders, there is no such
expectation.