instr: Improve handling of errors related to program counter
Document that we only support a single register as the program
counter currently. This prevents some instruction sets from being
modeled, for example 8086 and IA-32.

Check not only that the program counter exists, but also that it is
a single register. If no supported program counter exists, continue
parsing but skip analysis that depends on the program counter.