disasm: Prepare filler objects for completing a CodeTemplate
This is inspired by the way instruction decoding is done: using
processing objects that are put into a data structure. Those can
then be optimized before the actual mass processing begins.

Even in this early stage, there is already a significant speedup.
Although to be fair, most of that speedup originates from not running
the simplifier for value placeholders without a code block, which
could also have been done without introducing processing objects.