instr: Use dedicated code for creating encoding namespace
Sharing the data structure with the semantical use of the context is
not possible, since a match placeholder can have an encoding type
that is different from its semantical type.

Sharing code for creating the namespace is possible and was
previously done using _buildPlaceholder(). However, the only part
of _buildPlaceholder() that is still relevant is already available
as reusable code in ContextNamespace.

I added a specific error message for when a placeholder with a
computed value occurs in the encoding field: a generic "unknown name"
message would be confusing, since the name does exist in the context
but is omitted from the encoding namespace.