instr: Create separate code blocks for each computed value placeholder
Previously, a single builder was used and each placeholder code block
had all preceding computed placeholder's computations effectively
inlined. This is not something I want the analyzer to rely on if at
some point in the future modes will be created in other ways than
by the parser.

But if the analyzer can't rely on it, then it is clearer to have the
parser produce isolated code blocks. So now a new builder is used for
each placeholder.