instr: Don't skip entire mode line on placeholder value error
Since errors in the value computation are unlikely to cause further
errors down the line, there is no need to abort parsing of the line.

We do still abort parsing if there is a naming error, since the
context is parsed for both semantics and encoding and bad names
would be reported twice otherwise.