instr: Put the parsing of encoding element in separate function
There is only one caller, but that caller's code is unweildy because
there are so many error cases that need to be checked and reported.
Splitting some of it off into functions makes it easier to read.

The exception handling around emitLoad() was removed, because that
method cannot actually raise BadInput.