instr: Put auxiliary encoding width check in separate function
All information needed for this check is available in EncodingExpr
and EncodingMultiMatch, so there is no need to mix the check with
the parsing. Separating the check makes the code easier to read.

Unlike before, we now only perform the width check if the encoding
fields were all parsed without errors. This should avoid error spam.