instr: Move check for missing placeholders into separate function
The main difference is that we now build the sets of present
placeholders from the encoding expressions instead of from the
parse nodes. This might be useful in the future, if we want to
apply the same verification to definitions that do not come
from a .instr file.

Also the flow of error handling was changed to reduce the chance
of error spam: we only check for missing placeholders if there
were no errors parsing of the encoding field.