instr: Remove per-instruction operand role analysis
Some operand roles can be determined by looking at a single
instruction, but others require tracing a value's usage through
multiple instructions. For example, in "ld hl,$c023 ; ld (hl),a"
the operand $c023 is a data address, but this cannot be determined
by looking at any of the instructions in isolation.

Since we have to perform the multi-instruction analysis anyway,
doing the single-instruction analysis as well is unnecessary.

Note that we're going to do this analysis on instruction instances
in the future, not on instruction definitions. Therefore I removed
the 'roles' property from the Placeholder class.