instr: Add ModeMatch class to combine all aspects of a match
Encoding is newly implemented, mnemonic was moved from iterMnemonic()
and semantics aren't implemented yet. CodeTemplate.buildMatch() can
act as inspiration for the semantics, but the conversion is not so
straightforward, so I'd rather save it for a later commit.

One of the motivations for introducing this new class is that it is
useful to bring together all aspects in the same place: iterMnemonic()
was already using the 'values' dictionary produced by buildMatch(), so
there is overlap between the three aspects. Also both iterMnemonic()
and buildMatch() were using EncodeMatch.

Another motivation is to flatten the match: instead of recursively
processing each aspect separately, we can create a ModeMatch for
each submode match. Some of these submode matches are candidates
for caching in the future.