instr: Remove restriction that program counter must be single register
This requirement originated from the per-instruction analysis, which
was removed. When I re-introduce analysis of writes to the program
counter, that analysis might work only for a single-register program
counter at first. But that would be a limitation on the quality of
the analysis, not a limitation on instruction set definitions.