instr: Update 'reg' block syntax and parser
The register definition syntax was one of the first things I designed
and it was no longer consistent with things that were designed later,
such as function bodies. I did keep the ability to omit the type
beyond the first register on a line, but the rest of the syntax is
now consistent with functions.

The expression parser is now used for the full line within 'reg'
blocks, while previously it was only used for alias values. As a
result, we now have span information on the locations of register
definitions.

In the Z80 definition, I decided that since the shadow registers are
never accessed individually, it was more efficient to model them as
16-bit registers.