instr: Add support for prefix definitions
The current design is based on the needs of the Z80; it may need to be
reworked later to support other instruction sets. The only other one
I found so far is x86; prefixes seem to be rare.

As of this commit, the prefix definitions are documented, added in the
Z80 instr file and parsed. They are not yet stored in a data structure
or applied when (dis)assembling.