instr: Reject use of registers and I/O channels in encodings
This was already indirectly rejected by the code that determines the
decode mapping, but we can give more specific error messages if we
reject these cases earlier.

Originally I wanted to pass a namespace without registers and I/O
channels in it, but being able to use global functions and constants
could be useful.