instr: Clean up encoding decomposition code
Document _decomposeBitString(), return first element as BitString
instead of having two very different types (Storage and Expression).
Use consistent names of indices.

The "unsupported storage type" error message should be impossible
for the end user to trigger, since _parseEncodingExpr() already
rejects unsupported storage types.

Instead of logging each error, decomposeEncoding() now raises
BadInput on the first error. While in theory this limits our ability
to report as many errors as possible, in practice I don't think the
kind of errors this routine will catch will happen very often. The
motivation for this change is that by not having to pass the reader,
it becomes easier to re-use this routine, for example for prefixes.