instr: Fix subtle bug in optimization attempt in inlineBlock()
This bug was introduced in 43b9cb9a (instr: Add fast path in
inlineBlock() where no substitution takes place).

A test case was added to demonstrate the bug.

Loads are identified by the Load node object itself. So if a Load
object is added to more than one node list, it is seen as the exact
same load occurring at multiple moments in time, which is incorrect.

For the moment I removed the fast path. Maybe I'll want to add
a correct fast path at some point in the future, but right now
I'm not focusing on this part of the code.