instr: Assign unique names to placeholders within a match
In the parsed definitions names are only unique within one mode
entry, so if submode match is performed, that submode might use
names that also occur in a higher level entry, or in entries
matched by a different submode.

This will make it a lot easier to implement new optimizations on
decode search trees in the future.