instr: Fill in submode matches that are not represented in the encoding
Some modes, such as the Z80 "hlx" mode, match entirely based on
decode flags. So when a decode tree is instantiated for a particular
combination of flags, submode matches like "hlx" can be replaced
by the guaranteed match.

There are still some limitations in the filling process, marked with
TODOs in this commit. The semantics TODO is not triggered by any of
the existing instruction set definitions. The no-placeholders TODO
is not an issue yet since an entry from a MatchFoundDecoder will not
contain any placeholders.