instr: z80: Fix order of mode entries for 'res' and 'set'
The order didn't matter back when 'reg8' didn't include an entry with
encoding %110, but in 3 weeks ago I decided to add '(hl)' in 'reg8'
and '(hl)' is not a valid value for the 3rd operand in the unofficial
variant of these instructions. The entry for the official instruction
must override the entry for the unofficial instruction and therefore
be placed last.