instr: Remove method for logging at critical level
This method was never called.

Whether to abort processing on an error is determined not by the
severity of the error itself, but by how likely it is that further
errors that will be detected are genuine errors vs fallout of earlier
errors. Also, the decision to continue or not is typically made at
the end of processing a chunk of data, not immediately when an error
is reported.