instr, asm: Use InputLocation not just for error reporting
Since every input text has to be accompanied by an InputLocation
object during processing so that we can produce good error reports,
I figured it might be simpler to pass only the location objects
around and fetch the input text from those where needed.

Apart from needing fewer arguments and variables, it is also possible
to wrap the regex operations so that they produce new location
objects. This means that the location information for matches within
matches is now correct from the start: no more need to shift it
afterwards if the text being searched might not be on the start of
the line.