instr: Fix handling of using a non-existing function return value
This problem would be detected by Expression.checkScalar(), but that
just raises a ValueError that is not intended to be caught.

The fix revolves around adding a parsing category for statements that
are valid in the semantics field of mode tables: this is a subset of
statements valid in function bodies, but a superset of expressions.
In this category function calls are handled explicitly and when no
value is returned that is not considered an error, unlike function
calls that are part of an expression, which must produce a value.