instr: Remove support for 'unit' as a non-value expression
Since the previous commit removed the only use of the 'unit' object,
this commit removes that object, its class (Unit) and the support
code in the Expression base class for rejecting unit expressions
where they are not allowed.

Originally I thought it would be simpler to allow expressions to
contain function calls that produce no value and trigger errors
when and where those non-value expressions are used as values.
But it turns out that is more complex and error-prone than just
having two different pieces of code to deal with function calls.