instr: Prevent error spam when a definition's value contains errors
Previously, if a parse error occurred on the right hand side of
a definition, the name for that definition was not added to the
namespace. This would then very likely lead to an unknown name
being reported later. To avoid that, we now insert a dummy
reference into the namespace.