instr: Prevent error spam on calls to function that contains errors
When a function body contains errors, a Function object is created
with None for its code block. Previously, when a call to such a
function was inlined, it was considered to not return anything.
However, this can trigger an error in the caller if a value is
required at the place of the call. Now we instead consider such
a function to return a dummy bit string (BadBits) that matches
its declared return type.

I also fixed the docstring of inlineFunctionCall() that stated that
it returns a Reference, while it returns a BitString instead. I guess
it did return a Reference at some point in the past.