asm: Turn Token class into Tokenizer
Tokens are typically consumed in order, so having support for that
in the 'tokens' module is useful. Random access is still possible
by explicitly storing the iteration results in a sequence.

The 'end' token was replaced by a property to check for end-of-input.
Having a sentinel is useful for implementing peek() and eat(), but
it's not necessary to expose this.

The 'whitespace' token was removed; the tokenizer now automatically
skips whitespace using a hardcoded pattern.

The 'comment' token is no longer discarded early in the asm parser.
Parsing comments will be useful to extract API documentation.