Raise ValueError for non-matched groups in InputLocation regexes
Returning None for non-matched groups led to a lot of unnecessary
asserts when using the match object for regular expressions that can
only match all groups or not match at all.

For expressions that can optionally match a group, the new hasGroup()
method can be used to test whether a group was matched, prior to
getting the matched value.