instr: Make LocalNamespace.createCodeBlock() more strict
Previously, any name for the returned reference that didn't exist in
the namespace was interpreted as the block not returning anything.
This is too likely to hide bugs: if a name should exist but does not,
the inconsistency will not be spotted until the generated code block
gets used.

In particular, this means that 'checkdef' does not spot the bug;
we have to for example disassemble code that uses a particular
mode entry in order to trigger the bug.

The solution is to require the caller to explicitly pass None if the
block is not supposed to return anything and raise an exception if
a non-existing name is passed.