instr: Pass returned reference to LocalNamespace.createCodeBlock()
This goes a step beyond the previous commit by making the returned
reference an argument to createCodeBlock(). Most callers already had
the reference available, so passing that is more explicit than looking
up the reference by name.

This also fixes a bug with values computed in the context, such as
relative addresses, which are not given a namespace entry in the
placeholder namespace and therefore couldn't be looked up there
by name.