instr: Clean up error reporting for redefined imported namespace elements
There is no point in storing the None locations; we can just do the
lookup with get() instead.

The check for the new location being None is not necessary, since the
argument type guarantees it is never None.

We now explicitly state that an imported name is redefined if that
is the case, which should make the error message more meaningful.