decode: Implement Encoding.fillPlaceholders()
The problem I had before annotating the substitution function (see
removed TODO) was because that function was incorrect: it tried to
return an EncodeExpr where a BitString was expected.

This bug wasn't triggered because at the moment we only use the fill
to optimize no-encoding matches in the decode tree and if there is
no (empty) encoding there is also nothing to substitute in.