instr: Don't create a CodeTemplate without semantics
A mode entry's semantics can be missing because of an error in the
definition file or because we're skipping semantics for performance.
In those case, instead of having a template with a None code block,
we now have a None template in the mode entry.

This simplifies the template implementation, while not adding much
complexity in the mode entry. It also avoids unnecessary overhead
when specializing mode entries (filling in placeholders etc.).