instr: Move EncodeMatch into 'mode' module
The string variant of EncodeMatch was used in isinstance(), while it
is only suitable for type checking.

I think the class can actually be used to for mode matches in general,
not just encoding matches: if a match is made on mnemonics, the match
result would not be any different. So I can see it getting merged with
or replacing the ModeMatch class at some point.