instr: Make placeholder values constants in semantics field
Until now, they were emitted as value arguments, which means they
acted like variables rather than constants. This does not match the
syntax used to define them: a variable should be defined using the
':=' operator instead of '='.

I think that in practice it is not desirable to attempt to modify
placeholders, so the constants interpretation would be better.