Define Segment as start + width, instead of start + end
I think we use width slightly more often than end index in code that
can be converted to use Segments.

I considered this for the instruction set definition language as well,
but found it too confusing to have a Python-like syntax for slicing
with different semantics and couldn't find another syntax that I liked.