Fix nested slice decomposition
The decomposition flattening tests had so much functionality in their
test helper function that they were effectively testing the helper
without ever calling the code that should have been tested.

After making the test call the actual `BitString.decompose()`, the
test failed and it turned out this was a bug in the decomposition
of nested slices. The outer slice needs to be shifted to the
coordinate system of the base string before it can be clipped
to the inner slice.