Use "raise ... from" to chain or replace exception
As a rule of thumb, if the new exception re-uses the message string
from the old exception (`f"{ex}"`) I picked chaining (`from ex`) and
if it doesn't I picked replacing (`from None`).

Both options are superior to not using the `from` clause, in which
case exceptions are chained but the backtrace printer cannot tell
a deliberately chained exception apart from an unexpected exception
that occurred in the `except` clause.