Add Invoke task that looks for unused code
The tool `vulture` is used for this. There are a lot of false positives
at the moment, so I'm not sure how useful it is, but I think it's worth
keeping the command around for a while and see if it can be made more
useful in the future, for example by decorating the `binfmt` classes.