Move flagsRequired from ParsedModeEntry to ModeEntry
The required prefix flags must be known to produce a full encoded
instruction from a `ModeMatch`, but this information was only present
in `ParsedModeEntry`, which is used in instruction decoding but not
available when encoding.

Since `ParsedModeEntry` has the `ModeEntry` as a field, it doesn't
lose access to `flagsRequired` in this move.