Do not return mutable set as AbstractSet
For sequences and mappings, it is relatively safe to return
a mutable object via an immutable interface. However, in the
case of sets the `|=` operator is valid on `AbstractSet` but
will modify the set if applied to a mutable set.

I also updated some purely internal uses of immutable sets to
use `frozenset[]` instead of `AbstractSet`, as we don't need
API flexibility there and being explicit is more clear.