Remove old code that substitutes labels for integer literals
This needs to be re-implemented in a processing step before formatting,
both because we only need it during disassembly and not other uses of
formatting and because we need the tracing results to be able to tell
apart code addresses, data addresses and other literals.

I don't think this old code will be very valuable when implementing
this feature properly and removing it simplifies the formatting code.