Add test case for disassembling instructions with non-unique opcode
Some instructions can be encoded in multiple ways, but only one way
can be picked by the assembler. Therefore if we want to do disassembly
that produces an identical image after a round trip, non-preferred
opcodes must be disassembled to data directives.

Since mode matching on mnemonics isn't implemented yet, there is no
simple fix for this, so I'm committing the failing test for now.