Reject negative segment width as well on construction
Instead of overriding `__new__()`, the range checks are now performed
in the `__post_init__()` hook. In principle I prefer to check before
instantiating an object, but in practice this is likely to cause fewer
troubles than messing with `__new__()`.

A new unit test checks the negative offset and width rejection.