Decode full instruction length even if there is ultimately no match
When a real CPU encounters an unknown instruction, it will decode
it in the same way as known instructions with similar opcodes. In
particular, every item it fetches from memory while decoding will
advance the program counter.

We had a test case that checked for the realistic behavior already,
but until now it failed.