Move looked_ahead computation into a fetcher wrapper class
Not all fetchers need this functionality.

This also fixes an issue where the look ahead size is reported wrong
when prefixes are present: the prefix decoding code advances the
fetcher, which creates a new object that didn't report lookups back
to the original fetcher. The new wrapper fixes that behavior and
as a result one of the expected test results changed.

The `update()` method may seem redundant at first sight, but the usual
`fetcher = fetcher.advance(steps)` approach would create an enormous
chain of fetcher wrappers when disassembling larger sections of code.