Make locations on ParseNodes optional
We want to be able to construct `ParseNode` instances when doing code
generation, when there is no source file.

In particular, the ability to reformat code requires the pretty printer
to operate on `ParseNodes` rather than `Expressions`, as we lose close
resemblance to the original code in the conversion to `Expressions`.
But we also need to pretty print disassembly output, where we don't
have a source file.

Maybe `ParseNode` isn't the right name when the parser is not the only
code creating them, but that is something that can be fixed later.