Inline `BasicBlock.verify()` method
This is for internal consistency checks; it doesn't need to be part
of the public interface of `BasicBlock`.

It could have been turned into a private method instead, but I want
to get rid of the subclassing in the codeblock simplifier and that
is easier when there is no method at all.