Argument fetchers are no longer allowed to return `None`
The documented behavior was that if the fetcher function returned
`None`, the code inliner would preserve the original argument.
However, that is error prone: the context in which the block is
being inlined may not have an argument of the same name, or maybe
it does exist but it has a different meaning and/or width.

The error handling was made a bit more detailed as well.