Remove caching from `BasicBlock.arguments`
We determine all the arguments on block construction so we can reject
name clashes early. Therefore, the on-demand nature of `const_property`
is not useful here.

An additional advantage is that we can keep the raising of `ValueError`
out of the public interface, as it never happens when the user fetches
this property, because construction of the block would have failed.