Clean up `TestNamespace.parent`
One code block calling another is not the same as nested blocks:
"outer" and "inner" refers to the call graph, not namespaces.
So instead of inheriting a namespace, both blocks share the same
global namespace.

Even in the case of nested blocks, the inner block would inherit
the outer block's namespace, not vice versa.