Disable "statement has no effect" warning for now
This started generating false positives when the local variable tracing
was moved into the builder, because no more nodes would be emitted when
a statement only changed local variables.

I'll have to decide whether to abandon this warning, make it work with
local variable tracing or revert the change that put tracing into the
builder.