Only treat undefined local variables as a problem if they persist
Instead of treating the read of the undefined variable as an error,
we now trace the undefined value and report an error if that value
makes it into a storage, I/O index or returned value.

The main motivation for this is that once branches will be added,
it is no longer an issue if a local variable is not initialized in
a particular basic block, because another basic block might have
initialized it.