Add a test to demonstrate that undefined values can be eliminated
I'm not sure yet whether this behavior is useful or just an accidental
property of the algorithm, but it still makes sense to verify that the
code acts as I expect it does.

Also fixed a bug in the docstring parser, which would skip the last
code line if no unindented text followed it.