Trace stored values in code block builder
The values of local variables were already traced by the builder,
now we also trace the values of storages there.

This simplifies the code a bit, since loads and stores are already
separated and we don't need to patch up the returned bitstrings.
And as we're not modifying existing nodes, it might make it possible
to turn nodes into immutable objects.