Replace node abstract base class by sum type
We actually support only `Load` and `Store`, not any other subclass
of `AccessNode` that the user might define. By making `AccessNode`
a sum type, we can express that in the type system.

To avoid duplicated boilerplate code, I turned `Load` and `Store`
into dataclasses, which eliminates almost all of the boilerplate.