Perform message formatting before calling `ErrorCollector` methods
In the logging system, it makes sense to not format a message until
it is determined that it will actually reach the log. However, with
`ErrorCollector` there won't be many messages to log and the majority
of the potential calls are for errors, which will always be formatted.

One advantage of doing formatting before the call is that f-strings
can be used. Another advantage is that it avoids problems where
accidentally a formatted string is passed as the pattern, which can
break if the formatted string happens to contains placeholders.