Store mask rather than width inside `BadValue`
When we're going to use `BadValue` for simplified expressions, the
result of applying an operator to a `BadValue` is likely another
`BadValue`, but with a mask equal to the operator's computed mask.
These masks might carry more information than just a width.

The reason for storing a mask (instead of hardcoding it to -1) is
that when an undefined value is ultimately not used, we don't have
to consider it an error. This is consistent with how we treat
undefined local variables.