Remove distribution of AND over OR and vice versa
This added a lot of complexity and wasn't all that useful: only
a single test case failed without these mechanisms.

I added a simpler distribution rule that masks OR terms that are
irrelevant to the final value due to the OR literal. This made
the previously mentioned test case pass again and also supports
a new test case that didn't pass with the previous code.