Document mnemonic definition syntax design
This design is not implemented yet and will likely require tweaking
when implementing it.

I'm adding more structure (restrictions) to the mnemonic syntax,
for two main reasons:
- to allow better error detection and reporting
- to simplify the API that works with mnemonics

The main change is that we now assume comma-separated operands.
In practice, every architecture I've seen uses this, but the current
definition and assembly parsers don't require it, which makes our
parser output an unstructured mess.