Drop claim that context can contain references
None of the currently supported instruction sets use references in
the context. In fact, support for them was never implemented.

Also I'm not sure what a reference would even mean when used in the
encoding or mnemonic column. References are useful for defining
semantics, but as both functions and submodes support references,
we don't require reference support in the context as well.