Avoid unnecessary nested `try`
We handle errors in the context by moving on to the next line, so there
is nothing more to handle after that.

This is for clarity only: the outer `try` has a `pass` statement as its
handler, so it didn't do anything anyway. Its only purpose is to skip
the creation of data structures if the parsing didn't succeed.