Separate builder from namespace
I can see the appeal of passing only one object around instead of two,
but that is not enough of a reason to put the builder inside of the
namespace.

Maybe it makes sense to have a builder that contains a namespace
instead, but I want the separation to be archived in any case.