Simplify syntax for defining return type
Instead of having the return type in the function header, it is now
part of the declaration of `ret`. This avoids a bunch of special case
handling in the parser. It also makes the syntax more consistent.

The down side is that it's no longer obvious what (if anything)
a function returns by looking just at its header. However, as this
is not a general purpose programming language and our functions
tend to be pretty small, I don't see this as an issue in practice.