Restrict declaration part of definition to const/ref
Variable are not allowed in definitions: instead they are declared
and then assigned to, as their value is only valid until the next
assignment.

The parser was written with this restriction in mind, but the type
annotations didn't reflect it, so `assert isinstance()` statements
were used to cover the distance.