Skip body when function arguments are rejected
Since we consider argument error fatal, we don't parse the block.
But that means we must skip the lines in the block or they will
cause syntax errors.

An alternative would be to consider argument errors non-fatal,
but it is quite likely we'd report bogus errors as fallout of
the argument not being defined.