Do not store global namespace in `InstructionSet`
The namespaces are designed for parsing. They're not ideal for use
later as they have imprecise type information: `NamespaceValue` is
a sum type.

For doing emulation it would be useful to list the registers and
I/O channels for an instruction set. Like the program counter, we
can extract those from the namespace in the parser and then pass
them with strict types to the `InstructionSet` constructor.