Fix rejection of mode match placeholders in context expressions
While this use was rejected before, it raised an exception in a way
that would not get caught and reported.

It took some thought to find a way to report this type of problem with
the location of the offending use attached, but I think that raising
an exception on load/store does the trick.