Handle integer literals in mnemonics via synthetic placeholders
This is the idea that triggered the whole refactoring of the context
parsing code. However, now that it is actually in place, I'm wondering
whether it might be simpler to have `FixedValueReference`s in the
mnemonic, not just for literals, but for value placeholders too.

In other words, I'm convinced unifying integer literals and value
placeholders will simplify things, but I'm less certain that the
`ValuePlaceholder` class is the way to do it.